grsec: Apply grsecurity-2.1.14-2.6.31.6-200911121812.patch
Signature-tree: f4f7cd3962510c7c27d1c0a119bcaafefcbaf587

+++ drivers/mfd/ab3100-core.c
@@ -465,7 +465,7 @@ static int ab3100_get_set_reg_open_file(struct inode *inode, struct file *file)
 	return 0;
 }
 
-static int ab3100_get_set_reg(struct file *file,
+static ssize_t ab3100_get_set_reg(struct file *file,
 			      const char __user *user_buf,
 			      size_t count, loff_t *ppos)
 {
